+++ zerver/views/streams.py
@@ -451,6 +451,7 @@ def add_subscriptions_backend(
         "subscriptions", json_validator=add_subscriptions_schema
     ),
     invite_only: bool = REQ(json_validator=check_bool, default=False),
+    is_web_public: bool = REQ(json_validator=check_bool, default=False),
     stream_post_policy: int = REQ(
         json_validator=check_int_in(Stream.STREAM_POST_POLICY_TYPES),
         default=Stream.STREAM_POST_POLICY_EVERYONE,
@@ -483,6 +484,7 @@ def add_subscriptions_backend(
             stream_dict_copy["description"] = stream_dict["description"].replace("\n", " ")
 
         stream_dict_copy["invite_only"] = invite_only
+        stream_dict_copy["is_web_public"] = is_web_public
         stream_dict_copy["stream_post_policy"] = stream_post_policy
         stream_dict_copy["history_public_to_subscribers"] = history_public_to_subscribers
         stream_dict_copy["message_retention_days"] = parse_message_retention_days(